Created the disabler.
This disabler is an Anton App exploit thats uses curl to make a learning list that is filled with to much data for anton to handle.
If it gets above 11000 Learn lists it will not load the page and the account will need longer to even just load the icon.

This is still meant for educational purposes and will probably will be fixed soon. 
Dont give your authtoken and more to strangers as they can and probably WILL login to your account or destroy it.

I am not responsible for what you use this tool for.